refactor: extract element-specific binding to ElementBinder
