+++ src/TemplateCollection.php
@@ -1,0 +1,13 @@
+<?php
+namespace Gt\DomTemplate;
+
+use Gt\Dom\Element;
+
+class TemplateCollection {
+	public function get(
+		Element $context,
+		?string $templateName = null
+	):TemplateElement {
+
+	}
+}
